# fix/Reintroduce `WHERE` filter to the Sparse Query for PgVectorStore