--- a/llama_index/vector_stores/postgres.py
+++ b/llama_index/vector_stores/postgres.py
@@ -472,10 +472,14 @@ class PGVectorStore(BasePydanticVectorStore):
         ts_query = func.plainto_tsquery(
             type_coerce(self.text_search_config, REGCONFIG), query_str
         )
-        stmt = select(  # type: ignore
-            self._table_class,
-            func.ts_rank(self._table_class.text_search_tsv, ts_query).label("rank"),
-        ).order_by(text("rank desc"))
+        stmt = (
+            select(  # type: ignore
+                self._table_class,
+                func.ts_rank(self._table_class.text_search_tsv, ts_query).label("rank"),
+            )
+            .where(self._table_class.text_search_tsv.op("@@")(ts_query))
+            .order_by(text("rank desc"))
+        )
 
         # type: ignore
         return self._apply_filters_and_limit(stmt, limit, metadata_filters)