@hatianzhang 2023-12-14
@rendyfebry can you ensure this revert does not reintroduce the issue mentioned
```
I think the hybrid search postgres query does not need the additional where clause, as it massively restricts the volume of returned rows. Mostly for my case, the number of returned rows is 0.

Ranking and limiting is already performed in the circumference of the query generation, which should be sufficient enough to get good results.
If the query stays as is, that basically means that every (normalised) word needs to be existent in the node/db row that is currently evaluated.

Even the current example (https://colab.research.google.com/github/run-llama/llama_index/blob/main/docs/examples/vector_stores/postgres.ipynb) fails for me when setting similarity_top_k to 0.
```