@rendyfebry 2023-12-14
@hatianzhang The first part already explained in the PR description.

As for this one, I don't even think that's an issue. This is rather an expectation.
> If the query stays as is, that basically means that every (normalised) word needs to be existent in the node/db row that is currently evaluated. 

For example, if your raw query is `russo ukraine war` the `plainto_tsquery()` to `russia & ukraine & war`, which means we will only search for nodes that mention all three of them, as mentioned by the previous PR. But I see that as an expectation, we definitely don't want to return the node about `vietnam war` just because it matches one of the words only right?