@rendyfebry 2023-12-14
@hatianzhang To be fair, there's a way to make that more flexible, instead of assuming the user wants to match all words or not, we should give them more flexibility. Instead of `plainto_tsquery()` we shouls use `to_tsquery()`

> plainto_tsquery takes every word in the search string and assumes that they should all be connected by an "and" operator ('&').

> to_tsquery() provides boolean abilities, but each word in the search string must be separated by a boolean operator. ('&' '|' or '!') (and, or, not).


But that would be out of this PR scoope, I'll leave that for future improvement.