@logan-markewich 2023-12-14
`we definitely don't want to return the node about vietnam war just because it matches one of the words only right?` -- I would argue we do, and this is what rerankers are for

Normal user queries are more like natural questions, I don't 100% understand how the text search works here, but does that mean every word in the question will have to match the retrieved nodes?

Fine with merging this for now I think? but this is probably be addressed in another PR :)