@rendyfebry 2023-12-15
@logan-markewich That's fair

But AFAIK for RAG in general, using the whole user prompt for retrieval query is too naive, and that applies to all retrievers, not only hybrid keyword searches. And that's the purpose of query transformation.

For example, if the prompt is **"What do you know about the Russo-Ukraine War?"**, we should use query transformation so our retrieval only queries with **"Russo-Ukraine War"** instead of the whole prompt.