@rendyfebry 2023-12-15
BTW @logan-markewich @hatianzhang to be clear, my main concern for this PR is on the performance impact of not using the filter. As we can see the difference is 2x on query execution time.

For other improvements, I'll leave that to the others.