@mroedder-d7 2023-12-18
@logan-markewich @rendyfebry @hatianzhang well ok, now we´re back at square one. 🙂 

For my specific use case (and possibly for thousands of others) I would argue that db performance is much less of a problem than being forced to send a query transformation to an LLM, which might take several seconds instead of an increase of 300ms which was shown by @rendyfebry above.  
DB performance optimization is a well-known problem and can be tackled from all sorts of angles. 

And mentioning again: please everybody try to run your project notebook that´s also used in the docs: https://colab.research.google.com/github/run-llama/llama_index/blob/main/docs/examples/vector_stores/postgres.ipynb

It will not work after this change.

Maybe it would be better to keep the examples functional to not mislead users and, if there´s a specific use case, to let the users design the query by themselves. Of course there´s always the option to create your own VectorStores. Which quite frankly, I will do now as I cannot trust the constantly changing functionality anymore.