@logan-markewich 2023-12-18
@mroedder-d7 my apologies. Let me figure out a fix that works for both issues here. Probably I can add a `strict hybrid` option or something like that