@mroedder-d7 2023-12-18
@logan-markewich no need to apologise, it´s all good 😄 
I would always recommend to keep things open for extension, because there will never be enough option toggles for all use cases. I built a subclass and overrode one function, now I have my perfect tailor-made solution :)